Add files via upload
This is the file of the data obtained in the experiment as a result of the pre test and post test, it has a modification of the original file to make the commitment not to reveal the information of the participants. Therefore, the "username" column was modified by "Type of user" where it is indicated if it is a member of the Cri or not.